refactor(Model tests): Refactor models to incorporate a storage strategy param and factories.
