+++ tests/test_permission_repository.py
@@ -1,0 +1,3 @@
+"""
+Tests for PermissionRepository.
+"""
